Move empty_dir_wrapper.go into the storage directory.
