+++ test/e2e/storage/vsphere_volume_diskformat.go
@@ -52,7 +52,7 @@ import (
 	11. Delete PVC, PV and Storage Class
 */
 
-var _ = SIGDescribe("Volume Disk Format [Volumes]", func() {
+var _ = SIGDescribe("Volume Disk Format", func() {
 	f := framework.NewDefaultFramework("volume-disk-format")
 	var (
 		client            clientset.Interface
